moved pkg/hostname to internal/hostname
