+++ internal/sampling/samplingstrategy/adaptive/aggregator.go
@@ -15,7 +15,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/sampling/samplingstrategy"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/samplingstore"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/samplingstore/model"
-	"github.com/jaegertracing/jaeger/pkg/hostname"
+	"github.com/jaegertracing/jaeger/internal/hostname"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 )
